Adjustments and bugfixes. Beta released.
Released version 0.9.0 beta.
Settings screen adjustments:  removed the apply button beause settings will be aplied real-time; removed the edit button for tcp destinations as it is just as easy to delete and add a new one; added a stylesheet for the settings screen containing styles that should not affect the main screen.
Modified styles for the main screen.
Added the README.md.
Added a packaged version of the applciation.